Make scripts into a module
